add read settings

+++ src/services/apiCabins.js
@@ -9,13 +9,26 @@ export async function getCabins() {
   }
   return data;
 }
-export async function createCabin(newCabin) {
-  const imgName = `${Math.random()}-${newCabin.image.name}`.replaceAll("/", "");
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imgName}`;
 
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
+  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
+    "/",
+    ""
+  );
+
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  // create/edit cabin
+  let query = supabase.from("cabins");
+  // create
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+  // edit
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
@@ -23,9 +36,10 @@ export async function createCabin(newCabin) {
   }
   //if the new cabin works,upload img
   // https://wnkflyhijxygezvcyhuj.supabase.co/storage/v1/object/public/cabin-images//cabin-001.jpg
+  if (hasImagePath) return data;
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
-    .upload(imgName, newCabin.image);
+    .upload(imageName, newCabin.image);
   //delete cabin if there was  an error with the img
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
